+++ src/GapicClientTrait.php
@@ -32,6 +32,7 @@
 
 namespace Google\ApiCore;
 
+use Google\ApiCore\Call;
 use Google\ApiCore\LongRunning\OperationsClient;
 use Google\ApiCore\Middleware\CredentialsWrapperMiddleware;
 use Google\ApiCore\Middleware\FixedHeaderMiddleware;
@@ -48,6 +49,7 @@ use Google\ApiCore\Options\ClientOptions;
 use Google\ApiCore\Options\TransportOptions;
 use Google\Auth\CredentialsLoader;
 use Google\Auth\FetchAuthTokenInterface;
+use Google\Auth\GetUniverseDomainInterface;
 use Google\LongRunning\Operation;
 use Google\Protobuf\Internal\Message;
 use Grpc\Gcp\ApiConfig;
@@ -75,6 +77,8 @@ trait GapicClientTrait
     private string $serviceName = '';
     private array $agentHeader = [];
     private array $descriptors = [];
+    /** @var array<callable> $middlewareCallables */
+    private array $middlewareCallables = [];
     private array $transportCallMethods = [
         Call::UNARY_CALL => 'startUnaryCall',
         Call::BIDI_STREAMING_CALL => 'startBidiStreamingCall',
@@ -82,6 +86,42 @@ trait GapicClientTrait
         Call::SERVER_STREAMING_CALL => 'startServerStreamingCall',
     ];
     private bool $isNewClient;
+
+    /**
+     * Add a middleware to the call stack by providing a callable which will be
+     * invoked at the start of each call, and will return an instance of
+     * {@see MiddlewareInterface} when invoked.
+     *
+     * The callable must have the following method signature:
+     *
+     *     callable(MiddlewareInterface): MiddlewareInterface
+     *
+     * An implementation may look something like this:
+     * ```
+     * $client->addMiddleware(function (MiddlewareInterface $handler) {
+     *     return new class ($handler) implements MiddlewareInterface {
+     *         public function __construct(private MiddlewareInterface $handler) {
+     *         }
+     *
+     *         public function __invoke(Call $call, array $options) {
+     *             // modify call and options (pre-request)
+     *             $response = ($this->handler)($call, $options);
+     *             // modify the response (post-request)
+     *             return $response;
+     *         }
+     *     };
+     * });
+     * ```
+     *
+     * @param callable $middlewareCallable A callable which returns an instance
+     *                 of {@see MiddlewareInterface} when invoked with a
+     *                 MiddlewareInterface instance as its first argument.
+     * @return void
+     */
+    public function addMiddleware(callable $middlewareCallable): void
+    {
+        $this->middlewareCallables[] = $middlewareCallable;
+    }
 
     /**
      * Initiates an orderly shutdown in which preexisting calls continue but new
@@ -167,6 +207,8 @@ trait GapicClientTrait
             'libVersion' => null,
             'apiEndpoint' => null,
             'clientCertSource' => null,
+            // if the universe domain hasn't been explicitly set, assume GDU ("googleapis.com")
+            'universeDomain' => GetUniverseDomainInterface::DEFAULT_UNIVERSE_DOMAIN,
         ];
 
         $supportedTransports = $this->supportedTransports();
@@ -180,6 +222,9 @@ trait GapicClientTrait
                 'stubOpts' => ['grpc.service_config_disable_resolution' => 1]
             ];
         }
+
+        // Keep track of the API Endpoint
+        $apiEndpoint = $options['apiEndpoint'] ?? null;
 
         // Merge defaults into $options starting from top level
         // variables, then going into deeper nesting, so that
@@ -195,42 +240,33 @@ trait GapicClientTrait
             $options['transportConfig']['rest'] += $defaultOptions['transportConfig']['rest'];
         }
 
-        $this->modifyClientOptions($options);
+        // These calls do not apply to "New Surface" clients.
+        if (!$this->isNewClientSurface()) {
+            $preModifiedOptions = $options;
+            $this->modifyClientOptions($options);
+            // NOTE: this is required to ensure backwards compatiblity with $options['apiEndpoint']
+            if ($options['apiEndpoint'] !== $preModifiedOptions['apiEndpoint']) {
+                $apiEndpoint = $options['apiEndpoint'];
+            }
 
-        // serviceAddress is now deprecated and acts as an alias for apiEndpoint
-        if (isset($options['serviceAddress'])) {
-            $options['apiEndpoint'] = $this->pluck('serviceAddress', $options, false);
+            // serviceAddress is now deprecated and acts as an alias for apiEndpoint
+            if (isset($options['serviceAddress'])) {
+                $apiEndpoint = $this->pluck('serviceAddress', $options, false);
+            }
+        } else {
+            // Ads is using this method in their new surface clients, so we need to call it.
+            // However, this method is not used anywhere else for the new surface clients
+            // @TODO: Remove this in GAX V2
+            $this->modifyClientOptions($options);
         }
-
-        // If an API endpoint is set, ensure the "audience" does not conflict
+        // If an API endpoint is different form the default, ensure the "audience" does not conflict
         // with the custom endpoint by setting "user defined" scopes.
-        if ($options['apiEndpoint'] != $defaultOptions['apiEndpoint']
+        if ($apiEndpoint
+            && $apiEndpoint != $defaultOptions['apiEndpoint']
             && empty($options['credentialsConfig']['scopes'])
             && !empty($options['credentialsConfig']['defaultScopes'])
         ) {
             $options['credentialsConfig']['scopes'] = $options['credentialsConfig']['defaultScopes'];
-        }
-
-        if (extension_loaded('sysvshm')
-                && isset($options['gcpApiConfigPath'])
-                && file_exists($options['gcpApiConfigPath'])
-                && isset($options['apiEndpoint'])) {
-            $grpcGcpConfig = self::initGrpcGcpConfig(
-                $options['apiEndpoint'],
-                $options['gcpApiConfigPath']
-            );
-
-            if (array_key_exists('stubOpts', $options['transportConfig']['grpc'])) {
-                $options['transportConfig']['grpc']['stubOpts'] += [
-                    'grpc_call_invoker' => $grpcGcpConfig->callInvoker()
-                ];
-            } else {
-                $options['transportConfig']['grpc'] += [
-                    'stubOpts' => [
-                        'grpc_call_invoker' => $grpcGcpConfig->callInvoker()
-                    ]
-                ];
-            }
         }
 
         // mTLS: detect and load the default clientCertSource if the environment variable
@@ -248,11 +284,54 @@ trait GapicClientTrait
 
         // mTLS: If no apiEndpoint has been supplied by the user, and either
         // GOOGLE_API_USE_MTLS_ENDPOINT tells us to, or mTLS is available, use the mTLS endpoint.
-        if ($options['apiEndpoint'] === $defaultOptions['apiEndpoint']
-            && $this->shouldUseMtlsEndpoint($options)
-        ) {
-            $options['apiEndpoint'] = self::determineMtlsEndpoint($options['apiEndpoint']);
+        if (is_null($apiEndpoint) && $this->shouldUseMtlsEndpoint($options)) {
+            $apiEndpoint = self::determineMtlsEndpoint($options['apiEndpoint']);
         }
+
+        // mTLS: It is not valid to configure mTLS outside of "googleapis.com" (yet)
+        if (isset($options['clientCertSource'])
+            && $options['universeDomain'] !== GetUniverseDomainInterface::DEFAULT_UNIVERSE_DOMAIN
+        ) {
+            throw new ValidationException(
+                'mTLS is not supported outside the "googleapis.com" universe'
+            );
+        }
+
+        if (is_null($apiEndpoint)) {
+            if (defined('self::SERVICE_ADDRESS_TEMPLATE')) {
+                // Derive the endpoint from the service address template and the universe domain
+                $apiEndpoint = str_replace(
+                    'UNIVERSE_DOMAIN',
+                    $options['universeDomain'],
+                    self::SERVICE_ADDRESS_TEMPLATE
+                );
+            } else {
+                // For older clients, the service address template does not exist. Use the default
+                // endpoint instead.
+                $apiEndpoint = $defaultOptions['apiEndpoint'];
+            }
+        }
+
+        if (extension_loaded('sysvshm')
+            && isset($options['gcpApiConfigPath'])
+            && file_exists($options['gcpApiConfigPath'])
+            && !empty($apiEndpoint)
+        ) {
+            $grpcGcpConfig = self::initGrpcGcpConfig(
+                $apiEndpoint,
+                $options['gcpApiConfigPath']
+            );
+
+            if (!array_key_exists('stubOpts', $options['transportConfig']['grpc'])) {
+                $options['transportConfig']['grpc']['stubOpts'] = [];
+            }
+
+            $options['transportConfig']['grpc']['stubOpts'] += [
+                'grpc_call_invoker' => $grpcGcpConfig->callInvoker()
+            ];
+        }
+
+        $options['apiEndpoint'] = $apiEndpoint;
 
         return $options;
     }
@@ -413,7 +492,8 @@ trait GapicClientTrait
 
         $this->credentialsWrapper = $this->createCredentialsWrapper(
             $options['credentials'],
-            $options['credentialsConfig']
+            $options['credentialsConfig'],
+            $options['universeDomain']
         );
 
         $transport = $options['transport'] ?: self::defaultTransport();
@@ -433,15 +513,15 @@ trait GapicClientTrait
      * @return CredentialsWrapper
      * @throws ValidationException
      */
-    private function createCredentialsWrapper($credentials, array $credentialsConfig)
+    private function createCredentialsWrapper($credentials, array $credentialsConfig, string $universeDomain)
     {
         if (is_null($credentials)) {
-            return CredentialsWrapper::build($credentialsConfig);
+            return CredentialsWrapper::build($credentialsConfig, $universeDomain);
         } elseif (is_string($credentials) || is_array($credentials)) {
-            return CredentialsWrapper::build(['keyFile' => $credentials] + $credentialsConfig);
+            return CredentialsWrapper::build(['keyFile' => $credentials] + $credentialsConfig, $universeDomain);
         } elseif ($credentials instanceof FetchAuthTokenInterface) {
             $authHttpHandler = $credentialsConfig['authHttpHandler'] ?? null;
-            return new CredentialsWrapper($credentials, $authHttpHandler);
+            return new CredentialsWrapper($credentials, $authHttpHandler, $universeDomain);
         } elseif ($credentials instanceof CredentialsWrapper) {
             return $credentials;
         } else {
@@ -791,6 +871,11 @@ trait GapicClientTrait
             'metadataReturnType'
         ]);
 
+        foreach (\array_reverse($this->middlewareCallables) as $fn) {
+            /** @var MiddlewareInterface $callStack */
+            $callStack = $fn($callStack);
+        }
+
         return $callStack;
     }
 
@@ -1068,6 +1153,7 @@ trait GapicClientTrait
      *
      * @param array $options
      * @access private
+     * @internal
      */
     protected function modifyClientOptions(array &$options)
     {