+++ src/Middleware/RetryMiddleware.php
@@ -40,7 +40,7 @@ use GuzzleHttp\Promise\PromiseInterface;
 /**
  * Middleware that adds retry functionality.
  */
-class RetryMiddleware
+class RetryMiddleware implements MiddlewareInterface
 {
     /** @var callable */
     private $nextHandler;
@@ -176,7 +176,7 @@ class RetryMiddleware
     private function getRetryFunction()
     {
         return $this->retrySettings->getRetryFunction() ??
-            function (\Exception $e, array $options): bool {
+            function (\Throwable $e, array $options): bool {
                 // This is the default retry behaviour, i.e. we don't retry an ApiException
                 // and for other exception types, we only retry when the error code is in
                 // the list of retryable error codes.